Refactor pr-checks to run as a Javascript Github Action directly
