+++ api/octokit.ts
@@ -11,16 +11,16 @@ import { graphql } from '@octokit/graphql'
 import { exec } from 'child_process'
 import type { GraphQlQueryResponseData } from '@octokit/graphql'
 import {
-	Comment,
-	GitHub,
-	GitHubIssue,
-	Issue,
-	Milestone,
-	ProjectAndColumnIds,
+	type Comment,
+	type GitHub,
+	type GitHubIssue,
+	type Issue,
+	type Milestone,
+	type ProjectAndColumnIds,
 	projectType,
-	PullRequest,
-	Query,
-	User,
+	type PullRequest,
+	type Query,
+	type User,
 } from './api'
 
 let numRequests = 0
