--- pr-checks/Check.js
@@ -1,7 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.Check = void 0;
-class Check {
-}
-exports.Check = Check;
-//# sourceMappingURL=Check.js.map
