+++ pr-checks/package.json
@@ -1,25 +1,11 @@
 {
-	"name": "grafana-github-actions",
-	"version": "1.0.0",
+	"name": "grafana-github-actions-pr-checks",
 	"private": true,
-	"description": "GitHub Actions used by Grafana for triaging issues",
+	"version": "1.0.0",
+	"type": "module",
 	"scripts": {
-		"test": "jest --silent",
-		"build": "tsc && yarn workspace grafana-github-actions-pr-checks build",
-		"lint": "eslint -c .eslintrc --fix --ext .ts .",
-		"watch": "tsc --watch"
+		"build": "rollup --config rollup.config.js"
 	},
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/grafana/grafana-github-actions.git"
-	},
-	"keywords": [],
-	"author": "",
-	"license": "MIT",
-	"bugs": {
-		"url": "https://github.com/grafana/grafana-github-actions/issues"
-	},
-	"homepage": "https://github.com/grafana/grafana-github-actions#readme",
 	"dependencies": {
 		"@actions/core": "^1.6.0",
 		"@actions/exec": "^1.1.0",
@@ -37,9 +23,14 @@
 		"lodash": "^4.17.20",
 		"lodash.escaperegexp": "^4.1.2",
 		"micromatch": "^4.0.2",
+		"rollup": "^4.60.1",
 		"semver": "^7.5.4"
 	},
 	"devDependencies": {
+		"@rollup/plugin-commonjs": "^29.0.2",
+		"@rollup/plugin-json": "^6.1.0",
+		"@rollup/plugin-node-resolve": "^16.0.3",
+		"@rollup/plugin-typescript": "^12.3.0",
 		"@types/chai": "^4.2.10",
 		"@types/jest": "^27.4.0",
 		"@types/node": "^20.9.2",
@@ -56,23 +47,7 @@
 		"prettier": "2.5.1",
 		"ts-jest": "^27.1.3",
 		"ts-node": "^8.10.2",
-		"typescript": "^4.7.2",
+		"typescript": "^6.0.2",
 		"yargs": "^17.7.2"
-	},
-	"packageManager": "yarn@4.11.0",
-	"husky": {
-		"hooks": {
-			"pre-commit": "npm run lint && npm run build && git add \"**/*.js\""
-		}
-	},
-	"engines": {
-		"node": ">=22",
-		"yarn": ">=3.1.1",
-		"npm": "please-use-yarn"
-	},
-	"workspaces": {
-		"packages": [
-			"pr-checks"
-		]
 	}
 }
