+++ pr-checks/src/Dispatcher.test.ts
@@ -1,14 +1,17 @@
 import { expect } from 'chai'
 import { Dispatcher } from './Dispatcher'
 import { Check } from './Check'
-import { CheckState, CheckSubscriber } from './types'
+import { CheckState, type CheckSubscriber } from './types'
 import { context } from '@actions/github'
+import { OctoKitIssue } from '../../api/octokit'
 
 describe('Dispatcher', () => {
+	const api = new OctoKitIssue('abc-123', context.repo, { number: 123 })
+
 	describe('isMatch and dispatch', () => {
 		describe('When check subscribes and webhook event unhandled is triggered', () => {
 			it('Should not dispatch to check', async () => {
-				const d = new Dispatcher(null)
+				const d = new Dispatcher(api)
 				const check = new TestCheck()
 				check.subscribe(d)
 				context.eventName = 'unhandled'
@@ -20,7 +23,7 @@ describe('Dispatcher', () => {
 
 		describe('When check subscribes and webhook event e1 is triggered', () => {
 			it('Should dispatch to check 2 times', async () => {
-				const d = new Dispatcher(null)
+				const d = new Dispatcher(api)
 				const check = new TestCheck()
 				check.subscribe(d)
 				context.eventName = 'e1'
@@ -35,7 +38,7 @@ describe('Dispatcher', () => {
 
 		describe('When check subscribes and webhook event e1 with action a1 is triggered', () => {
 			it('Should dispatch to check 6 times', async () => {
-				const d = new Dispatcher(null)
+				const d = new Dispatcher(api)
 				const check = new TestCheck()
 				check.subscribe(d)
 				context.eventName = 'e1'
@@ -50,7 +53,7 @@ describe('Dispatcher', () => {
 
 		describe('When check subscribes and webhook event e1 with action a2 is triggered', () => {
 			it('Should dispatch to check 4 times', async () => {
-				const d = new Dispatcher(null)
+				const d = new Dispatcher(api)
 				const check = new TestCheck()
 				check.subscribe(d)
 				context.eventName = 'e1'
@@ -65,7 +68,7 @@ describe('Dispatcher', () => {
 
 		describe('When check subscribes and webhook event e2 is triggered', () => {
 			it('Should dispatch to check 1 time', async () => {
-				const d = new Dispatcher(null)
+				const d = new Dispatcher(api)
 				const check = new TestCheck()
 				check.subscribe(d)
 				context.eventName = 'e2'
@@ -80,7 +83,7 @@ describe('Dispatcher', () => {
 
 		describe('When check subscribes and webhook event e2 with action a1 is triggered', () => {
 			it('Should dispatch to check 3 times', async () => {
-				const d = new Dispatcher(null)
+				const d = new Dispatcher(api)
 				const check = new TestCheck()
 				check.subscribe(d)
 				context.eventName = 'e2'
@@ -95,7 +98,7 @@ describe('Dispatcher', () => {
 
 		describe('When check subscribes and webhook event e2 with action a2 is triggered', () => {
 			it('Should dispatch to check 2 times', async () => {
-				const d = new Dispatcher(null)
+				const d = new Dispatcher(api)
 				const check = new TestCheck()
 				check.subscribe(d)
 				context.eventName = 'e2'
