+++ pr-checks/src/checks/MilestoneCheck.test.ts
@@ -1,7 +1,7 @@
 import { context } from '@actions/github'
-import { EventPayloads } from '@octokit/webhooks'
+import { type EventPayloads } from '@octokit/webhooks'
 import { expect } from 'chai'
-import { PullRequest } from '../../api/api'
+import { type PullRequest } from '../../../api/api'
 import { Dispatcher } from '../Dispatcher'
 import { CheckState } from '../types'
 import { MilestoneCheck } from './MilestoneCheck'
