+++ tsconfig.test.json
@@ -5,12 +5,7 @@
 		"moduleResolution": "node",
 		"resolveJsonModule": true,
 		"noEmit": true,
-		"lib": [
-			"es2020"
-		]
+		"lib": ["es2020"]
 	},
-	"include": [
-		"**/*.test.ts",
-		"__mocks__/**/*.ts"
-	]
+	"include": ["**/*.test.ts", "__mocks__/**/*.ts"]
 }
